# Conflicts:
#	src/main/java/com/book/book_project/controller/ProductController.java
#	src/main/java/com/book/book_project/service/ProductService.java
#	src/main/java/com/book/book_project/service/ProductServiceImpl.java
#	src/main/java/com/book/book_project/service/ReviewService.java
#	src/main/java/com/book/book_project/service/ReviewServiceImpl.java

+++ src/main/java/com/book/book_project/dto/ReviewInterface.java
@@ -6,10 +6,8 @@ import com.book.book_project.entity.ProductEntity;
 import java.sql.Timestamp;
 
 public interface ReviewInterface {
-	Long getReviewseq();
-
+	int getReviewseq();
 	ProductEntity getBookid();
-
 	MemberEntity getUserid();
 	String getReviewer();
 	String getReviewcontent();
